fix(nav): move nav highlighting checks from controller to container
