+++ src/openshift/components/CryostatController.tsx
@@ -39,61 +39,9 @@ class CryostatControllerComponent extends React.Component<CryostatControllerProp
 
   private loadCryostat = async (): Promise<void> => {
     await this.getCryostatConfig();
-    this.checkNavHighlighting();
     this.applyUIDefaults();
     this.setDocLayout();
     this.setState({ loaded: true });
-  };
-
-  /**
-   * At the moment, nested routes will cause multiple nav items to be highlighted in the Console main navigation.
-   *
-   * This means by default the Dashboard nav item (route /cryostat/) will be highlighted when visiting any of the
-   * Cryostat plugin pages (routes /cryostat/whatever/) in addition to highlighting the currently selected page.
-   *
-   * This function checks the current href, and if it is not a Dashboard page then it will remove the pf-m-current class
-   * from the <a> classList to remove the highlighting. Due to this manual intervention, we also need to ensure that the
-   * pf-m-current returns when routing back to a Dashboard page, as it is not added back automatically.
-   */
-  private checkNavHighlighting = (): void => {
-    const currentHref = window.location.href;
-    if (!this.isDashboardRoute(currentHref)) {
-      this.removeDashboardNavHighlighting();
-    } else {
-      this.addDashboardNavHighlighting();
-    }
-  };
-
-  private isDashboardRoute = (href: string): boolean => {
-    return href.endsWith('/cryostat') || href.endsWith('/cryostat/') || href.includes('d-solo');
-  };
-
-  // Fetches the currently highlighted nav items, and removes pf-m-current from the Dashboard link if found
-  private removeDashboardNavHighlighting = (): void => {
-    const currentlySelected = document.getElementsByClassName('pf-m-current');
-    for (let i = 0; i < currentlySelected.length; i++) {
-      const href = currentlySelected[i].getAttribute('href');
-      if (!href) {
-        continue;
-      }
-      if (this.isDashboardRoute(href)) {
-        currentlySelected[i].classList.remove('pf-m-current');
-      }
-    }
-  };
-
-  // Fetches the nav link elements, and adds pf-m-current back to the Dashboard link
-  private addDashboardNavHighlighting = (): void => {
-    const navLinks = document.getElementsByClassName('pf-v5-c-nav__link');
-    for (let i = 0; i < navLinks.length; i++) {
-      const href = navLinks[i].getAttribute('href');
-      if (!href) {
-        continue;
-      }
-      if (this.isDashboardRoute(href)) {
-        navLinks[i].classList.add('pf-m-current');
-      }
-    }
   };
 
   private getCryostatConfig = async (): Promise<void> => {
